+++ src/agent.schema.ts
@@ -2,15 +2,6 @@
 
 import { z } from 'zod';
 
-export const AgentBase = z.object({
-	name: z.string()
-		.describe('The name of the agent'),
-	tags: z.array(z.string())
-		.describe('The tags of the agent'),
-})
-	.describe('Base information of the agent');
-export type AgentBase = z.infer<typeof AgentBase>;
-
 export const AgentRegistration = z.object({
 	tenant_id: z.string()
 		.describe('The tenant ID of the agent'),
@@ -24,14 +15,45 @@ export const AgentRegistration = z.object({
 	.describe('The registration of the agent');
 export type AgentRegistration = z.infer<typeof AgentRegistration>;
 
-export const AgentMetadata = AgentRegistration.merge(z.object({
+export const AgentBase = z.object({
+	name: z.string()
+		.describe('The name of the agent'),
+	tags: z.array(z.string())
+		.describe('The tags of the agent'),
+})
+	.describe('Base information of the agent');
+export type AgentBase = z.infer<typeof AgentBase>;
+
+export const AgentMetadata = z.object({
+	tenant_id: z.string()
+		.describe('The tenant ID of the agent'),
+	device_id: z.string().uuid()
+		.describe('The UUID of the host device'),
+	agent_id: z.string().uuid()
+		.describe('The UUID of the agent'),
+	create_timestamp: z.string().datetime()  // ISO 8601
+		.describe('The timestamp of the agent creation'),
 	modify_timestamp: z.string().datetime()
 		.describe('The timestamp of the agent modification'),
 	is_deleted: z.boolean().default(false)
 		.describe('The flag of the agent deletion'),
-}))
+})
 	.describe('The metadata of the agent');
 export type AgentMetadata = z.infer<typeof AgentMetadata>;
+
+export const AgentStateMetadata = z.object({
+	create_timestamp: z.string().datetime()  // ISO 8601
+		.describe('The timestamp of the agent state creation'),
+	modify_timestamp: z.string().datetime()
+		.describe('The timestamp of the agent state modification'),
+	is_deleted: z.boolean().default(false)
+		.describe('The flag of the agent state deletion'),
+})
+	.describe('The metadata of the agent state');
+export type AgentStateMetadata = z.infer<typeof AgentStateMetadata>;
+
+export const AgentStatusMetadata = AgentStateMetadata;
+export type AgentStatusMetadata = z.infer<typeof AgentStatusMetadata>;
 
 export const AgentStateBase = z.object({
 	url: z.string().url().nullable()
@@ -48,47 +70,29 @@ export const AgentStateBase = z.object({
 })
 	.describe('The state of the agent');
 export type AgentStateBase = z.infer<typeof AgentStateBase>;
-
-export const AgentState = AgentStateBase.merge(AgentMetadata);
+export const AgentState = AgentStateBase.merge(AgentStateMetadata);
 export type AgentState = z.infer<typeof AgentState>;
 
 export const AgentStatusBase = z.object({
+	url: z.string().url().nullable()
+		.describe('The URL of the agent'),
 	detail: z.record(z.any()).nullable()
 		.describe('The detail of the agent status'),
 })
 	.describe('The status of the agent');
 export type AgentStatusBase = z.infer<typeof AgentStatusBase>;
-
-export const AgentStatus = AgentStatusBase.merge(AgentMetadata);
+export const AgentStatus = AgentStatusBase.merge(AgentStatusMetadata);
 export type AgentStatus = z.infer<typeof AgentStatus>;
 
 export const Agent = AgentBase.merge(AgentMetadata)
 	.merge(z.object({
-		desired_state: AgentState
-			.omit({
-				tenant_id: true,
-				device_id: true,
-				agent_id: true,
-				is_deleted: true,
-			})
+		desired_state: AgentStateBase.merge(AgentStateMetadata)
 			.nullable()
 			.describe('The desired state of the agent'),
-		runtime_state: AgentState
-			.omit({
-				tenant_id: true,
-				device_id: true,
-				agent_id: true,
-				is_deleted: true,
-			})
+		runtime_state: AgentStateBase.merge(AgentStateMetadata)
 			.nullable()
 			.describe('The runtime state of the agent'),
-		runtime_status: AgentStatus
-			.omit({
-				tenant_id: true,
-				device_id: true,
-				agent_id: true,
-				is_deleted: true,
-			})
+		runtime_status: AgentStatusBase.merge(AgentStatusMetadata)
 			.nullable()
 			.describe('The runtime status of the agent'),
 
@@ -102,9 +106,6 @@ const sqliteDateSchema = z.string().transform((date) => {
 });
 
 export const DbDtoToAgentState = z.object({
-	tenant_id: z.string().uuid(),
-	device_id: z.string().uuid(),
-	agent_id: z.string().uuid(),
 	url: z.string().nullable(),
 	pull_interval: z.number().nullable(),
 	push_interval: z.number().nullable(),
@@ -124,9 +125,7 @@ export const DbDtoToAgentState = z.object({
 	});
 
 export const DbDtoToAgentStatus = z.object({
-	tenant_id: z.string().uuid(),
-	device_id: z.string().uuid(),
-	agent_id: z.string().uuid(),
+	url: z.string().nullable(),
 	detail: z.string().nullable(),
 	create_timestamp: sqliteDateSchema,
 	modify_timestamp: sqliteDateSchema,
@@ -180,6 +179,7 @@ export const DbDtoToAgent = z.object({
 	desired_state_detail: z.string().nullable().optional(),
 	desired_state_create_timestamp: sqliteDateSchema.optional(),
 	desired_state_modify_timestamp: sqliteDateSchema.optional(),
+	desired_state_is_deleted: z.number().default(0),
 	runtime_state_url: z.string().nullable().optional(),
 	runtime_state_pull_interval: z.number().nullable().optional(),
 	runtime_state_push_interval: z.number().nullable().optional(),
@@ -188,12 +188,15 @@ export const DbDtoToAgent = z.object({
 	runtime_state_detail: z.string().nullable().optional(),
 	runtime_state_create_timestamp: sqliteDateSchema.optional(),
 	runtime_state_modify_timestamp: sqliteDateSchema.optional(),
+	runtime_state_is_deleted: z.number().default(0),
+	runtime_status_url: z.string().nullable().optional(),
 	runtime_status_detail: z.string().nullable().optional(),
 	runtime_status_create_timestamp: sqliteDateSchema.optional(),
 	runtime_status_modify_timestamp: sqliteDateSchema.optional(),
+	runtime_status_is_deleted: z.number().default(0),
 })
 	.transform((dto): Agent => {
-		const desired_state = (
+		const desired_state: AgentState | null = (
 			   typeof dto.desired_state_url === "undefined"
 			&& typeof dto.desired_state_pull_interval === "undefined"
 			&& typeof dto.desired_state_push_interval === "undefined"
@@ -213,8 +216,9 @@ export const DbDtoToAgent = z.object({
 				: null,
 			create_timestamp: z.string().parse(dto.desired_state_create_timestamp),
 			modify_timestamp: z.string().parse(dto.desired_state_modify_timestamp),
+			is_deleted: Boolean(dto.desired_state_is_deleted),
 		};
-		const runtime_state = (
+		const runtime_state: AgentState | null = (
 			   typeof dto.runtime_state_url === "undefined"
 			&& typeof dto.runtime_state_pull_interval === "undefined"
 			&& typeof dto.runtime_state_push_interval === "undefined"
@@ -234,17 +238,21 @@ export const DbDtoToAgent = z.object({
 				: null,
 			create_timestamp: z.string().parse(dto.runtime_state_create_timestamp),
 			modify_timestamp: z.string().parse(dto.runtime_state_modify_timestamp),
+			is_deleted: Boolean(dto.runtime_state_is_deleted),
 		};
-		const runtime_status = (
-			   typeof dto.runtime_status_detail === "undefined"
+		const runtime_status: AgentStatus | null = (
+			   typeof dto.runtime_state_url === "undefined"
+			&& typeof dto.runtime_status_detail === "undefined"
 			&& typeof dto.runtime_status_create_timestamp === "undefined"
 			&& typeof dto.runtime_status_modify_timestamp === "undefined"
 		) ? null : {
+			url: dto.runtime_state_url ?? null,
 			detail: dto.runtime_status_detail
 				? JSON.parse(dto.runtime_status_detail)
 				: null,
 			create_timestamp: z.string().parse(dto.runtime_status_create_timestamp),
 			modify_timestamp: z.string().parse(dto.runtime_status_modify_timestamp),
+			is_deleted: Boolean(dto.runtime_status_is_deleted),
 		};
 		return {
 			// AgentBase