+++ src/device.schema.ts
@@ -2,15 +2,7 @@
 
 import { z } from 'zod';
 import { CoolReport } from './cool.schema.js';
-
-export const DeviceBase = z.object({
-	name: z.string()
-		.describe('The name of the device'),
-	tags: z.array(z.string())
-		.describe('The tags of the device'),
-})
-	.describe('Base information of the device, which is used to create a device');
-export type DeviceBase = z.infer<typeof DeviceBase>;
+import { ScreenDetails } from './screen-details.schema.js';
 
 export const DeviceRegistration = z.object({
 	tenant_id: z.string()
@@ -23,16 +15,47 @@ export const DeviceRegistration = z.object({
 	.describe('The registration of the device');
 export type DeviceRegistration = z.infer<typeof DeviceRegistration>;
 
-export const DeviceMetadata = DeviceRegistration.merge(z.object({
+export const DeviceBase = z.object({
+	name: z.string()
+		.describe('The name of the device'),
+	tags: z.array(z.string())
+		.describe('The tags of the device'),
+})
+	.describe('Base information of the device, which is used to create a device');
+export type DeviceBase = z.infer<typeof DeviceBase>;
+
+export const DeviceMetadata = z.object({
+	tenant_id: z.string()
+		.describe('The tenant ID of the device'),
+	device_id: z.string().uuid()
+		.describe('The UUID of the device'),
+	create_timestamp: z.string().datetime()  // ISO 8601
+		.describe('The timestamp of the device creation'),
 	modify_timestamp: z.string().datetime()
 		.describe('The timestamp of the device modification'),
 	is_deleted: z.boolean().default(false)
 		.describe('The flag of the device deletion'),
-}))
+})
 	.describe('The metadata of the device');
 export type DeviceMetadata = z.infer<typeof DeviceMetadata>;
 
+export const DeviceStateMetadata = z.object({
+	create_timestamp: z.string().datetime()  // ISO 8601
+		.describe('The timestamp of the device state creation'),
+	modify_timestamp: z.string().datetime()
+		.describe('The timestamp of the device state modification'),
+	is_deleted: z.boolean().default(false)
+		.describe('The flag of the device state deletion'),
+})
+	.describe('The metadata of the device state');
+export type DeviceStateMetadata = z.infer<typeof DeviceStateMetadata>;
+
+export const DeviceStatusMetadata = DeviceStateMetadata;
+export type DeviceStatusMetadata = z.infer<typeof DeviceStatusMetadata>;
+
 export const DeviceStateBase = z.object({
+	url: z.string().url().nullable()
+		.describe('The URL of the device'),
 	pull_interval: z.number().nullable()
 		.describe('The pull interval of the device'),
 	push_interval: z.number().nullable()
@@ -46,45 +69,12 @@ export const DeviceStateBase = z.object({
 })
 	.describe('The state of the device');
 export type DeviceStateBase = z.infer<typeof DeviceStateBase>;
-
-export const DeviceState = DeviceStateBase.merge(DeviceMetadata);
+export const DeviceState = DeviceStateBase.merge(DeviceStateMetadata);
 export type DeviceState = z.infer<typeof DeviceState>;
 
-export const ScreenOrientation = z.object({
-	angle: z.number()
-		.describe('The angle of the screen orientation'),
-	type: z.string()
-		.describe('The type of the screen orientation'),
-})
-	.describe('The screen orientation');
-export type ScreenOrientation = z.infer<typeof ScreenOrientation>;
-
-export const ScreenDetailed = z.object({
-	label: z.string()
-		.describe('The label of the screen details'),
-	left: z.number()
-		.describe('The left of the screen details'),
-	top: z.number()
-		.describe('The top of the screen details'),
-	width: z.number()
-		.describe('The width of the screen details'),
-	height: z.number()
-		.describe('The height of the screen details'),
-	devicePixelRatio: z.number()
-		.describe('The device pixel ratio of the screen details'),
-	orientation: ScreenOrientation
-		.describe('The orientation of the screen details'),
-})
-	.describe('The screen details');
-export type ScreenDetailed = z.infer<typeof ScreenDetailed>;
-
-export const ScreenDetails = z.object({
-	screens: z.array(ScreenDetailed),
-})
-	.describe('Details of all screens available to the device');
-export type ScreenDetails = z.infer<typeof ScreenDetails>;
-
 export const DeviceStatusBase = z.object({
+	url: z.string().url().nullable()
+		.describe('The URL of the device'),
 	user_agent: z.string().nullable()
 		.describe('The user agent of the device'),
 	device_memory: z.number().nullable()
@@ -104,34 +94,18 @@ export const DeviceStatusBase = z.object({
 })
 	.describe('The runtime status of the device');
 export type DeviceStatusBase = z.infer<typeof DeviceStatusBase>;
-
-export const DeviceStatus = DeviceStatusBase.merge(DeviceMetadata);
+export const DeviceStatus = DeviceStatusBase.merge(DeviceStatusMetadata);
 export type DeviceStatus = z.infer<typeof DeviceStatus>;
 
 export const Device = DeviceBase.merge(DeviceMetadata)
 	.merge(z.object({
-		desired_state: DeviceState
-			.omit({
-				tenant_id: true,
-				device_id: true,
-				is_deleted: true,
-			})
+		desired_state: DeviceStateBase.merge(DeviceStateMetadata)
 			.nullable()
 			.describe('The desired state of the device'),
-		runtime_state: DeviceState
-			.omit({
-				tenant_id: true,
-				device_id: true,
-				is_deleted: true,
-			})
+		runtime_state: DeviceStateBase.merge(DeviceStateMetadata)
 			.nullable()
 			.describe('The runtime state of the device'),
-		runtime_status: DeviceStatus
-			.omit({
-				tenant_id: true,
-				device_id: true,
-				is_deleted: true,
-			})
+		runtime_status: DeviceStatusBase.merge(DeviceStatusMetadata)
 			.nullable()
 			.describe('The runtime status of the device'),
 	}));
@@ -144,8 +118,7 @@ const sqliteDateSchema = z.string().transform((date) => {
 });
 
 export const DbDtoToDeviceState = z.object({
-	tenant_id: z.string().uuid(),
-	device_id: z.string().uuid(),
+	url: z.string().nullable(),
 	pull_interval: z.number().nullable(),
 	push_interval: z.number().nullable(),
 	min_backoff_interval: z.number().nullable(),
@@ -164,8 +137,7 @@ export const DbDtoToDeviceState = z.object({
 	});
 
 export const DbDtoToDeviceStatus = z.object({
-	tenant_id: z.string().uuid(),
-	device_id: z.string().uuid(),
+	url: z.string().nullable(),
 	user_agent: z.string().nullable(),
 	device_memory: z.number().nullable(),
 	hardware_concurrency: z.number().nullable(),
@@ -213,13 +185,14 @@ export const DbDtoToDeviceBase = z.object({
 		};
 	});
 export const DbDtoToDevice = z.object({
-	device_id: z.string().uuid(),
 	tenant_id: z.string().uuid(),
+	device_id: z.string().uuid(),
 	name: z.string(),
 	tags: z.string(),
 	create_timestamp: sqliteDateSchema,
 	modify_timestamp: sqliteDateSchema,
 	is_deleted: z.number().default(0),
+	desired_state_url: z.string().nullable().optional(),
 	desired_state_pull_interval: z.number().nullable().optional(),
 	desired_state_push_interval: z.number().nullable().optional(),
 	desired_state_min_backoff_interval: z.number().nullable().optional(),
@@ -227,6 +200,8 @@ export const DbDtoToDevice = z.object({
 	desired_state_agent_ids: z.string().nullable().optional(),
 	desired_state_create_timestamp: sqliteDateSchema.optional(),
 	desired_state_modify_timestamp: sqliteDateSchema.optional(),
+	desired_state_is_deleted: z.number().default(0),
+	runtime_state_url: z.string().nullable().optional(),
 	runtime_state_pull_interval: z.number().nullable().optional(),
 	runtime_state_push_interval: z.number().nullable().optional(),
 	runtime_state_min_backoff_interval: z.number().nullable().optional(),
@@ -234,6 +209,8 @@ export const DbDtoToDevice = z.object({
 	runtime_state_agent_ids: z.string().nullable().optional(),
 	runtime_state_create_timestamp: sqliteDateSchema.optional(),
 	runtime_state_modify_timestamp: sqliteDateSchema.optional(),
+	runtime_state_is_deleted: z.number().default(0),
+	runtime_status_url: z.string().nullable().optional(),
 	runtime_status_user_agent: z.string().nullable().optional(),
 	runtime_status_device_memory: z.number().nullable().optional(),
 	runtime_status_hardware_concurrency: z.number().nullable().optional(),
@@ -245,10 +222,12 @@ export const DbDtoToDevice = z.object({
 	runtime_status_error_stack: z.string().nullable().optional(),
 	runtime_status_create_timestamp: sqliteDateSchema.optional(),
 	runtime_status_modify_timestamp: sqliteDateSchema.optional(),
+	runtime_status_is_deleted: z.number().default(0),
 })
 	.transform((dto): Device => {
-		const desired_state = (
-			   typeof dto.desired_state_pull_interval === "undefined"
+		const desired_state: DeviceState | null = (
+			   typeof dto.desired_state_url === "undefined"
+			&& typeof dto.desired_state_pull_interval === "undefined"
 			&& typeof dto.desired_state_push_interval === "undefined"
 			&& typeof dto.desired_state_min_backoff_interval === "undefined"
 			&& typeof dto.desired_state_max_backoff_interval === "undefined"
@@ -256,6 +235,7 @@ export const DbDtoToDevice = z.object({
 			&& typeof dto.desired_state_create_timestamp === "undefined"
 			&& typeof dto.desired_state_modify_timestamp === "undefined"
 		) ? null : {
+			url: dto.desired_state_url ?? null,
 			pull_interval: dto.desired_state_pull_interval ?? null,
 			push_interval: dto.desired_state_push_interval ?? null,
 			min_backoff_interval: dto.desired_state_min_backoff_interval ?? null,
@@ -265,9 +245,11 @@ export const DbDtoToDevice = z.object({
 				: null,
 			create_timestamp: z.string().parse(dto.desired_state_create_timestamp),
 			modify_timestamp: z.string().parse(dto.desired_state_modify_timestamp),
+			is_deleted: Boolean(dto.desired_state_is_deleted),
 		};
-		const runtime_state = (
-			   typeof dto.runtime_state_pull_interval === "undefined"
+		const runtime_state: DeviceState | null = (
+			   typeof dto.runtime_state_url === "undefined"
+			&& typeof dto.runtime_state_pull_interval === "undefined"
 			&& typeof dto.runtime_state_push_interval === "undefined"
 			&& typeof dto.runtime_state_min_backoff_interval === "undefined"
 			&& typeof dto.runtime_state_max_backoff_interval === "undefined"
@@ -275,6 +257,7 @@ export const DbDtoToDevice = z.object({
 			&& typeof dto.runtime_state_create_timestamp === "undefined"
 			&& typeof dto.runtime_state_modify_timestamp === "undefined"
 		) ? null : {
+			url: dto.runtime_state_url ?? null,
 			pull_interval: dto.runtime_state_pull_interval ?? null,
 			push_interval: dto.runtime_state_push_interval ?? null,
 			min_backoff_interval: dto.runtime_state_min_backoff_interval ?? null,
@@ -284,9 +267,11 @@ export const DbDtoToDevice = z.object({
 				: null,
 			create_timestamp: z.string().parse(dto.runtime_state_create_timestamp),
 			modify_timestamp: z.string().parse(dto.runtime_state_modify_timestamp),
+			is_deleted: Boolean(dto.runtime_status_is_deleted),
 		};
-		const runtime_status = (
-			   typeof dto.runtime_status_user_agent === "undefined"
+		const runtime_status: DeviceStatus | null = (
+			   typeof dto.runtime_state_url === "undefined"
+			&& typeof dto.runtime_status_user_agent === "undefined"
 			&& typeof dto.runtime_status_device_memory === "undefined"
 			&& typeof dto.runtime_status_hardware_concurrency === "undefined"
 			&& typeof dto.runtime_status_vendor_webgl === "undefined"
@@ -298,6 +283,7 @@ export const DbDtoToDevice = z.object({
 			&& typeof dto.runtime_status_create_timestamp === "undefined"
 			&& typeof dto.runtime_status_modify_timestamp === "undefined"
 		) ? null : {
+			url: dto.runtime_state_url ?? null,
 			user_agent: dto.runtime_status_user_agent ?? null,
 			device_memory: dto.runtime_status_device_memory ?? null,
 			hardware_concurrency: dto.runtime_status_hardware_concurrency ?? null,
@@ -313,6 +299,7 @@ export const DbDtoToDevice = z.object({
 			error_stack: dto.runtime_status_error_stack ?? null,
 			create_timestamp: z.string().parse(dto.runtime_status_create_timestamp),
 			modify_timestamp: z.string().parse(dto.runtime_status_modify_timestamp),
+			is_deleted: Boolean(dto.runtime_status_is_deleted),
 		};
 		return {
 			// DeviceBase