+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## v4.0.2
+- Introduce new `AgentStateMetadata`, `AgentStatusMetadata`, `DeviceStateMetadata`, and `DeviceStatusMetadata`.
+- Replace `type` in `detail` with `url` in _state_ or _status_ objects.
+
 ## v3.2.1
 - Add `DeviceRegistration` as response to _device registration_ endpoint.
 - Add `AgentRegistration` as response to _create agent_ endpoint.